+++ CHANGELOG.md
@@ -3,11 +3,13 @@
 ## `jupyter_starters 0.2.0a0`
 
 - add notebook as starter [#13][]
+- add copying files and commands while starter is continuing [#18][]
 
 ## `@deathbeds/jupyterlab-starters 0.2.0a0`
 
 - add notebook metadata authoring [#13][]
 - add Jupyter markdown to `title`, `description` and `ui:help` in schema forms [#17][]
+- all starters launch in right sidebar [#18][]
 
 ## `jupyter_starters 0.1.0a3`
 
@@ -36,3 +38,4 @@
 
 [#13]: https://github.com/deathbeds/jupyterlab-starters/pull/13
 [#17]: https://github.com/deathbeds/jupyterlab-starters/pull/17
+[#18]: https://github.com/deathbeds/jupyterlab-starters/pull/18